refactor(optics): rename geolens_pkg mixins and fix pre-existing test failures
Rename mixin files for consistent naming:
- seidel.py → eval_seidel.py
- tolerance.py → eval_tolerance.py
- surf_ops.py → optim_ops.py
- view_3d.py → vis3d.py

Move match_materials() from geolens.py to optim_ops.py. Update all
imports, docstrings, and documentation references.

Fix three pre-existing test issues:
- test_material_default: expect "air" (Material normalizes None to "air")
- test_aspheric_higher_order: expect ai4*r^4 (ai list starts at a4, not a2)
- test_read_zmx: use available ef35mm_f2.0.zmx fixture
